refactor: 컨트롤러 분리
게임 진행과 관련된 요청 처리를 GameController로 그 이외의 요청에 대한 처리는 HomeController로 책임을 분리하였습니다.
