+++ src/main/java/chess/controller/GameController.java
@@ -13,34 +13,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
-public class SpringController {
+public class GameController {
 
     private final ChessService chessService;
 
-    public SpringController(ChessService chessService) {
+    public GameController(ChessService chessService) {
         this.chessService = chessService;
-    }
-
-    @GetMapping("/")
-    public String home(Model model, @RequestParam(value = "error", required = false) String error) {
-        List<String> gameNames = chessService.findAllGameName();
-        model.addAttribute("gameNames", gameNames);
-        model.addAttribute("error", error);
-        return "index";
-    }
-
-    @GetMapping("/load")
-    public String restart(@RequestParam("game_name") String gameName) {
-        Long id = chessService.findIdByGameName(gameName);
-
-        return "redirect:/game/" + id;
-    }
-
-    @PostMapping("/start")
-    public String start(@RequestParam("game_name") String gameName, @RequestParam("password") String password) {
-        Long id = chessService.save(gameName, password);
-
-        return "redirect:/game/" + id;
     }
 
     @GetMapping("/game/{id}")
@@ -94,19 +72,6 @@ public class SpringController {
         model.addAttribute("id", id);
 
         return "chess";
-    }
-
-    @PostMapping("/delete")
-    public String delete(@RequestParam(value = "game_name", required = false) String gameName,
-                         @RequestParam(value = "password", required = false) String password) {
-        chessService.deleteBy(gameName, password);
-
-        return "redirect:/";
-    }
-
-    @GetMapping("/error")
-    public String error() {
-        return "error";
     }
 
     private String makeCommand(String from, String to) {
